requirements.txt
langgraph
langchain_community
langchain
python-dotenv
ipykernel
langchain-openai
langgraph-cli
langgraph-cli[inmem]
fastmcp
langchain-mcp-adapters
google-generativeai
langgraph-cli
langgraph-cli[inmen]
fastmcp
langchain-mcp-adapters
